+++ req/req_test.go
@@ -452,7 +452,7 @@ func (s *ReqSuite) TestEncoding(c *C) {
 	}.Do()
 
 	c.Assert(err, NotNil)
-	c.Assert(err, ErrorMatches, `Can't encode request body \(json: unsupported type: func\(\)\)`)
+	c.Assert(err.Error(), Equals, "Can't encode request body: json: unsupported type: func()")
 	c.Assert(resp, IsNil)
 }
 
@@ -460,40 +460,32 @@ func (s *ReqSuite) TestRequestErrors(c *C) {
 	resp, err := Request{}.Do()
 
 	c.Assert(err, NotNil)
-	c.Assert(err, ErrorMatches, `Can't create request struct \(URL property can't be empty and must be set\)`)
+	c.Assert(err.Error(), Equals, "URL property can't be empty and must be set")
 	c.Assert(resp, IsNil)
 
 	resp, err = Request{URL: "ABCD"}.Do()
 
 	c.Assert(err, NotNil)
-	c.Assert(err, ErrorMatches, `Can't create request struct \(Unsupported scheme in URL\)`)
+	c.Assert(err.Error(), Equals, "Unsupported scheme in URL")
 	c.Assert(resp, IsNil)
 
 	resp, err = Request{URL: "http://127.0.0.1:60000"}.Do()
 
 	c.Assert(err, NotNil)
-	c.Assert(err, ErrorMatches, `Can't send request \(Get \"http://127.0.0.1:60000\": dial tcp 127.0.0.1:60000: connect: connection refused\)`)
+	c.Assert(err.Error(), Equals, `Can't send request: Get "http://127.0.0.1:60000": dial tcp 127.0.0.1:60000: connect: connection refused`)
 	c.Assert(resp, IsNil)
 
 	resp, err = Request{URL: "%gh&%ij"}.Do()
 
 	c.Assert(err, NotNil)
-	c.Assert(err, ErrorMatches, `Can't create request struct \(Unsupported scheme in URL\)`)
+	c.Assert(err.Error(), Equals, "Unsupported scheme in URL")
 	c.Assert(resp, IsNil)
 
 	resp, err = Request{Method: "ЩУП", URL: "http://127.0.0.1"}.Do()
 
 	c.Assert(err, NotNil)
-	c.Assert(err, ErrorMatches, `Can't create request struct \(net/http: invalid method "ЩУП"\)`)
+	c.Assert(err.Error(), Equals, `Can't create request: net/http: invalid method "ЩУП"`)
 	c.Assert(resp, IsNil)
-
-	e1 := RequestError{ERROR_BODY_ENCODE, "Test 1"}
-	e2 := RequestError{ERROR_CREATE_REQUEST, "Test 2"}
-	e3 := RequestError{ERROR_SEND_REQUEST, "Test 3"}
-
-	c.Assert(e1.Error(), Equals, "Can't encode request body (Test 1)")
-	c.Assert(e2.Error(), Equals, "Can't create request struct (Test 2)")
-	c.Assert(e3.Error(), Equals, "Can't send request (Test 3)")
 }
 
 func (s *ReqSuite) TestEngineInit(c *C) {
@@ -512,7 +504,7 @@ func (s *ReqSuite) TestEngineErrors(c *C) {
 	resp, err := eng.Do(Request{URL: "https://essentialkaos.com"})
 
 	c.Assert(err, NotNil)
-	c.Assert(err, ErrorMatches, `Can't create request struct \(Engine is nil\)`)
+	c.Assert(err, Equals, ErrNilEngine)
 	c.Assert(resp, IsNil)
 
 	eng = &Engine{}
@@ -523,7 +515,7 @@ func (s *ReqSuite) TestEngineErrors(c *C) {
 	resp, err = eng.Do(Request{URL: "https://essentialkaos.com"})
 
 	c.Assert(err, NotNil)
-	c.Assert(err, ErrorMatches, `Can't create request struct \(Engine.Dialer is nil\)`)
+	c.Assert(err, Equals, ErrNilDialer)
 	c.Assert(resp, IsNil)
 
 	eng = &Engine{}
@@ -533,7 +525,7 @@ func (s *ReqSuite) TestEngineErrors(c *C) {
 	resp, err = eng.Do(Request{URL: "https://essentialkaos.com"})
 
 	c.Assert(err, NotNil)
-	c.Assert(err, ErrorMatches, `Can't create request struct \(Engine.Transport is nil\)`)
+	c.Assert(err, Equals, ErrNilTransport)
 	c.Assert(resp, IsNil)
 
 	eng = &Engine{}
@@ -544,7 +536,7 @@ func (s *ReqSuite) TestEngineErrors(c *C) {
 	resp, err = eng.Do(Request{URL: "https://essentialkaos.com"})
 
 	c.Assert(err, NotNil)
-	c.Assert(err, ErrorMatches, `Can't create request struct \(Engine.Client is nil\)`)
+	c.Assert(err, Equals, ErrNilClient)
 	c.Assert(resp, IsNil)
 }